Cleaned up duplicate code a bit
